+++ src/Pages/Home.jsx
@@ -7,7 +7,7 @@ import Services from '../Components/Services'
 import LogoGrid from '../Components/LogoGrid'
 import EmpoweredPerformance from '../Components/empoweredPerformance'
 import TeamSolutions from '../Components/teamSolutions'
-import AboutUs from '../Components/AboutUs'
+import AboutUs from '../Components/aboutsection'
 import KeyDifferentiators from '../Components/KeyDifferetiators'
 import LeaderShipCoachingAndDevelopment from '../Components/LeaderShipCoachingAndDevelopment'
 import IndustriesWeServed from '../Components/Industries-We-Served'
